Add files via upload
On 09-01-2017 Constantijn Sikkel, the database specialist at ETCBC, gave a presentation about the feature `txt` (text type), (which is a little bit different than `domain`). With his permission I share the presentation with you. It clarifies some of the choices that are made when a researcher analyses a (B)H text for this specific feature.